+++ app/Controllers/ProyectosController.php
@@ -48,6 +48,15 @@ class ProyectosController extends BaseController
     }
 
     public function EditAction($categoria){
+        session_start();
+        $data = "";
+        $elementos = explode('/', $categoria);
+        $categ = end($elementos);
+        $data = Proyectos::getInstancia()->getUserId($categ);
+        if(!isset($_SESSION['id']) || $_SESSION['id'] != $data[0]['usuarios_id']){
+            header('Location: /');
+        }
+
         if(isset($_POST['editar'])){
             $data = "";
             $elementos = explode('/', $categoria);
@@ -91,12 +100,20 @@ class ProyectosController extends BaseController
     }
 
     public function DelAction($categoria){
+        session_start();
+        $data = "";
         $elementos = explode('/', $categoria);
         $categ = end($elementos);
+        $data = Proyectos::getInstancia()->getUserId($categ);
+        
+        if(!isset($_SESSION['id']) || $_SESSION['id'] == null || $_SESSION['id'] != $data[0]['usuarios_id']){
+            header('Location: /');
+            exit();
+        }
+        
         $proyecto = Proyectos::getInstancia();
         $proyecto->delete($categ);
 
-        // Llamamos a la función renderHTML
         header('Location: /portfolio/');
     }
 }
