+++ app/Controllers/SkillController.php
@@ -49,6 +49,14 @@ class SkillController extends BaseController
     }
 
     public function EditAction($categoria){
+        session_start();
+        $data = "";
+        $elementos = explode('/', $categoria);
+        $categ = end($elementos);
+        $data = SkillsUsuario::getInstancia()->getUserId($categ);
+        if(!isset($_SESSION['id']) || $_SESSION['id'] != $data[0]['usuarios_id']){
+            header('Location: /');
+        }
         if(isset($_POST['editar'])){
             $data = "";
             $elementos = explode('/', $categoria);
@@ -86,8 +94,16 @@ class SkillController extends BaseController
     }
 
     public function DelAction($categoria){
+        session_start();
+        $data = "";
         $elementos = explode('/', $categoria);
         $categ = end($elementos);
+        $data = SkillsUsuario::getInstancia()->getUserId($categ);
+        
+        if(!isset($_SESSION['id']) || $_SESSION['id'] == null || $_SESSION['id'] != $data[0]['usuarios_id']){
+            header('Location: /');
+            exit();
+        }
         $tarea = SkillsUsuario::getInstancia();
         $tarea->delete($categ);
